Splits code for modules
